An AI A* Application with Python
The project contains a data site consisting of the races of the day and the data is drawn from the Jockey Club of Turkey site. In this drawn dataset, each horse in the race includes the rank of each horse in the last six races. Each race matrix in the program is reduced to rows based on the average of the last six races of the horses, resulting in an overall matrix. Then, with the A* Search Algorithm, the horses that can win the race are predicted. Races with bank results as a result of the selected horses are given as output.

The races held on 22 and 29 May 2022 at Istanbul Veliefendi Racecourse are used as test data for this project.